html/rendering/non-replaced-elements/lists/lists-presentational-hints-ascii-case-insensitive.html

<!DOCTYPE html>
<meta charset="utf-8">
<link rel="help" href="https://html.spec.whatwg.org/#lists:presentational-hints">
<link rel="help" href="https://drafts.csswg.org/selectors-4/#attribute-case">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<ul type="circle"><li type="disc"></ul>
<ul type="circle"><li type="DiSc"></ul>
<ul type="circle"><li type="diſc"></ul>
<ul type="circle"><li type="square"></ul>
<ul type="circle"><li type="SqUaRe"></ul>
<ul type="circle"><li type="ſquare"></ul>
<script>
test(() => {
  const li = document.querySelectorAll("li");

  assert_equals(getComputedStyle(li[0]).getPropertyValue("list-style-type"), "disc");
  assert_equals(getComputedStyle(li[1]).getPropertyValue("list-style-type"), "disc");
  assert_equals(getComputedStyle(li[2]).getPropertyValue("list-style-type"), "circle");
  assert_equals(getComputedStyle(li[3]).getPropertyValue("list-style-type"), "square");
  assert_equals(getComputedStyle(li[4]).getPropertyValue("list-style-type"), "square");
  assert_equals(getComputedStyle(li[5]).getPropertyValue("list-style-type"), "circle");
});
</script>
